# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type Article struct {
	ID                int64
	OutletID          int64
	EventID           pgtype.Int8
	Url               string
	Headline          string
	Body              string
	PublishedAt       time.Time
	FirstSeenAt       time.Time
	ContentHash       string
	Minhash           []int32
	ReprintOfID       pgtype.Int8
	Development       string
	HappenedOn        pgtype.Date
	DateIsApproximate bool
	StepArticleID     pgtype.Int8
	LinkConfidence    pgtype.Float4
}

type ArticleEntity struct {
	ArticleID int64
	EntityID  int64
}

type Entity struct {
	ID            int64
	CanonicalName string
	Kind          string
	Aliases       []string
	SearchText    string
}

type Event struct {
	ID           int64
	Title        string
	FirstSeenAt  time.Time
	UpdatedAt    time.Time
	TimelineText string
	Search       interface{}
}

type EventEntity struct {
	EventID  int64
	EntityID int64
}

type Outlet struct {
	ID           int64
	Slug         string
	Name         string
	Domain       string
	IsAggregator bool
	Enabled      bool
	CrawlConfig  []byte
	CreatedAt    time.Time
	Coverage     string
}

type Summary struct {
	ArticleID     int64
	Text          string
	Recaps        []string
	Model         string
	PromptVersion string
	CreatedAt     time.Time
}